requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
pydub==0.25.1
openai==1.51.2
aiofiles==24.1.0
